+++ packages/core/src/List.svelte
@@ -1,6 +1,10 @@
 <script>
-	import ListActions from './ListActions.svelte'
-	import ListItems from './ListItems.svelte'
+	// import { navigable } from './actions'
+	import { createEventDispatcher } from 'svelte'
+	import { defaultFields } from './constants'
+	import { Text } from './items'
+
+	const dispatch = createEventDispatcher()
 
 	let className = 'list'
 	export { className as class }
@@ -8,41 +12,35 @@
 	export let fields = {}
 	export let using = {}
 	export let value = null
-	export let searchable = false
-	export let editable = false
 
-	let search
-	let filtered
-
-	function addItem() {
-		items = [...items, {}]
-		value = items[items.length - 1]
-	}
-	function deleteSelection() {
-		if (value) value.isDeleted = true
-	}
-	function clearSelection() {
-		value = null
+	function handleClick(item) {
+		value = item
+		dispatch('select', item)
 	}
 
-	$: filtered =
-		searchable && search && search.length
-			? items.filter((item) => item[fields.text].includes(search))
-			: items
+	$: fields = { ...defaultFields, ...fields }
+	$: using = { default: Text, ...using }
+	$: filtered = items.filter((item) => !item.isDeleted)
 </script>
 
-<list class="flex flex-col w-full {className}">
-	{#if searchable || editable}
-		<ListActions
-			bind:search
-			{searchable}
-			{editable}
-			on:delete={deleteSelection}
-			on:clear={clearSelection}
-			on:add={addItem}
-		/>
-	{/if}
-	<scroll class="flex flex-col h-full overflow-scroll">
-		<ListItems bind:items={filtered} {fields} {using} bind:value on:select />
-	</scroll>
+<list class="flex flex-col w-full flex-shrink-0 select-none {className}">
+	<slot />
+	{#each filtered as item}
+		{@const component = item[fields.component]
+			? using[item[fields.component]] || using.default
+			: using.default}
+		<!-- svelte-ignore a11y-click-events-have-key-events -->
+		<item
+			class="flex flex-shrink-0 flex-grow-0 min-h-8 items-center cursor-pointer w-full gap-2 select-none item"
+			class:is-selected={value === item}
+			on:click={() => handleClick(item)}
+		>
+			<svelte:component
+				this={component}
+				bind:content={item}
+				{fields}
+				on:change
+			/>
+		</item>
+	{/each}
 </list>