+++ packages/markdown/src/ComponentExamples.svelte
@@ -1,32 +1,33 @@
 <script>
-	import { CodeSnippet } from '@rokkit/markdown'
-	import PageNavigator from './PageNavigator.svelte'
-	import { defaultFields } from '@rokkit/core'
+	import CodeSnippet from './CodeSnippet.svelte'
+	import { defaultFields, PageNavigator } from '@rokkit/core'
+
 	let className = ''
 	export { className as class }
 
 	export let items = []
 	export let fields = {}
-	export let page = items.length ? items[0] : null
-
 	export let showCode = true
 
+	let page
+	$: name = (page?.component?.name ?? '').replace(/Proxy<(.*)>/, '$1')
+	$: page = !items.includes(page) ? items[0] : page
 	$: fields = { ...defaultFields, ...fields }
 </script>
 
-<page class="flex flex-col relative {className}">
-	<h1 class="text-xl font-bold">{page[fields.text]}</h1>
-	<p class="my-3 font-light">{page[fields.summary]}</p>
+<sample class="flex flex-col relative {className}">
+	<h1><span class="text-skin-600">{name}</span>: {page[fields.text]}</h1>
 	<content class="flex flex-row w-full h-full overflow-hidden">
 		<section
-			class="flex flex-col flex-grow items-center justify-center overflow-scroll"
+			class="flex flex-col flex-grow items-center p-4 justify-center overflow-scroll"
 		>
-			<wrapper class="flex flex-col self-center">
+			<wrapper class="flex flex-col h-full self-center {page.class}">
 				<svelte:component this={page.component} {...page.props} />
 			</wrapper>
+			<p>{page[fields.summary]}</p>
 		</section>
 		{#if showCode}
-			<pre class="flex flex-col h-full min-w-1/3 overflow-scroll">
+			<pre class="flex flex-col h-full min-w-2/5 overflow-scroll">
 				<CodeSnippet code={page.code} language="svelte" />
 			</pre>
 		{/if}
@@ -37,4 +38,4 @@
 		{/if}
 	</content>
 	<PageNavigator {items} {fields} bind:value={page} class="" />
-</page>
+</sample>