+++ sites/rokkit/src/lib/types.js
@@ -1,0 +1,19 @@
+/**
+ * @typedef Reference
+ * @property {string} source
+ * @property {string|Array.<string>} items
+ */
+
+/**
+ * @typedef Sample
+ * @property {string} title
+ * @property {string} code
+ * @property {string} summary
+ * @property {any} component
+ * @property {string} [class]
+ * @property {Array<Reference>} refs
+ * @property {Object<string, any>} props
+ * @property {Object<string, string>} [declarations]
+ */
+
+export default {}