+++ sites/rokkit/src/routes/(content)/[component]/examples/index.js
@@ -1,0 +1,10 @@
+import { default as list } from './list'
+import { default as accordion } from './accordion'
+
+/**
+ * @type {Object<string,Array<any>}
+ */
+export const data = {
+	list,
+	accordion
+}