+++ sites/rokkit/src/routes/(docs)/docs/+page.svelte
@@ -1,0 +1,16 @@
+<content class="flex gap-10 wrap p-10">
+	<a href="/list">
+		<h1>List</h1>
+		<img src="/preview/list.svg" alt="List Preview" />
+	</a>
+	<a href="/accordion">
+		<h1>Accordion</h1>
+		<img src="/preview/accordion.svg" alt="Accordion Preview" />
+	</a>
+</content>
+
+<style>
+	a {
+		@apply text-primary-50 text-lg;
+	}
+</style>