+++ sites/rokkit/src/routes/+layout.svelte
@@ -2,15 +2,16 @@
 	import 'uno.css'
 	import '../app.css'
 
-	import Background from './Background.svelte'
+	import Waves from './Waves.svelte'
 	import Header from './Header.svelte'
 	import { themable } from '@rokkit/core/actions'
 
 	export let data
 </script>
 
-<main use:themable class="flex flex-col w-full h-full relative">
-	<Background />
-	<Header menu={data.sections} />
+<svelte:body use:themable />
+<Header menu={data.sections} />
+<main class="flex flex-col w-full h-full relative">
 	<slot />
 </main>
+<Waves />