+++ packages/layout/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rokkit/layout
 
+## 1.0.0-next.1
+
+### Patch Changes
+
+- Improvements in accordion and list with examples.
+- Updated dependencies
+  - @rokkit/markdown@1.0.0-next.1
+  - @rokkit/core@1.0.0-next.1
+
 ## 1.0.0-next.0
 
 ### Patch Changes