+++ packages/markdown/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rokkit/markdown
 
+## 1.0.0-next.1
+
+### Patch Changes
+
+- Improvements in accordion and list with examples.
+- Updated dependencies
+  - @rokkit/core@1.0.0-next.1
+
 ## 1.0.0-next.0
 
 ### Patch Changes