+++ packages/themes/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rokkit/themes
 
+## 1.0.0-next.1
+
+### Patch Changes
+
+- Improvements in accordion and list with examples.
+
 ## 1.0.0-next.0
 
 ### Patch Changes